refactor(opencode): extract passive catalog normalization
